requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
openai==0.27.8
SQLAlchemy==2.0.5.post1
Flask-SQLAlchemy==3.0.2
python-dotenv==1.0.0
